requirements.txt
aiohttp==3.1.2
async-timeout==2.0.0
chardet==3.0.4
multidict==3.3.2
PyJWT==1.6.1
SQLAlchemy==1.2.0
yarl==0.16.0
